Test JSON decoding and fix implentation
Add a test that decodes a JSON object, and detect links in it. Then encode
it again and it must match the original file. The test failed (because the
link was a map[string]interface{} instead of a Node). The fix post-process
the result of the JSON decoder to convert map[string]interface{} to Node.