Test CBOR decoding and fix implementation
Same as d68f39de9a4acb036fcba7c9c6a8af2c17563b3a but for CBOR.

Add a test that decodes a CBOR object, and detect links in it. Then encode
it again and it must match the original file. The test failed (because the
link was a map[interface{}]interface{} instead of a Node). The fix
post-process the result of the CBOR decoder to convert
map[interface{}]interface{} to Node.

License: MIT
Signed-off-by: Mildred Ki'Lya <mildred-pub.git@mildred.fr>